산책_03 & 소비_04
-산책
등록, 수정, 날씨
완성

- 소비
페이지 정리 및 세부 기능 완성

+++ client/components/deletCnsm.jsx
@@ -3,7 +3,7 @@ import { Table, TableWrapper, Row, Rows, Col, Cols, Cell } from 'react-native-ta
 import { useContext, useEffect, useState } from "react";
 import { ConsumeContext } from "../contexts/consume-context";
 import Loading from "../customs/loading";
-import { StyleSheet, TouchableOpacity, View } from 'react-native';
+import { Alert, StyleSheet, TouchableOpacity, View } from 'react-native';
 import FontText from '../customs/fontText';
 import { deleteOneConsume, readConsumeAll } from '../api/consume';
 import { AppContext } from '../contexts/app-context';
@@ -16,28 +16,45 @@ function DeleteConsume() {
 
 
     useEffect(()=>{
-        // 소비내역 싹 불러오는 api 실행
-        const arr = consumeContext.data
-        // console.log(arr)
-        const combineArr = arr.map((one)=>{
-            // console.log(one)
-            return [one.date.slice(0,10), one.category, one.ammount, one.description, <TouchableOpacity onPress={()=>onDelete(one._id)}><FontText style={{textAlign :"center"}}>삭제</FontText></TouchableOpacity>]
-        }) 
-        // 실행해서 나온 값을 저장하게한다.
-        setData(combineArr)
+        if(consumeContext?.data){
+            // 소비내역 싹 불러오는 api 실행
+            const arr = consumeContext.data
+            // console.log(arr)
+            const combineArr = arr.map((one)=>{
+                // console.log(one)
+                return [<FontText style={styles.text}>{one.date.slice(0,10)}</FontText>,
+                <FontText style={styles.text}>{one.category}</FontText>,
+                <FontText style={styles.text}>{one.ammount}</FontText>, 
+                <FontText style={styles.text}>{one.description}</FontText>,
+                <TouchableOpacity onPress={()=>onDelete(one._id)}>
+                    <FontText style={[styles.text,{textAlign :"center"}]}>삭제</FontText>
+                </TouchableOpacity>]
+            }) 
+            // 실행해서 나온 값을 저장하게한다.
+            setData(combineArr)
+        }
     },[consumeContext.data])
 
     //func
     const onDelete = (item)=>{
+        Alert.alert(
+            "","소비내역을 삭제하시겠습니까?",[{
+                text : "취소"
+            },{
+                text : "삭제", onPress : ()=>{
+                    deleteOneConsume(item)
+                        .then((rcv)=>{
+                            // console.log(rcv);
+                            if(rcv.result){
+                                onRefresh();
+                            }
+                        })
+                        .catch(err=>console.log("onDelete => ",err))
+                }
+            }]
+        )
         // console.log(item)
-        deleteOneConsume(item)
-        .then((rcv)=>{
-            // console.log(rcv);
-            if(rcv.result){
-                onRefresh();
-            }
-        })
-        .catch(err=>console.log("onDelete => ",err))
+        
     }
 
     const onRefresh = ()=>{
@@ -51,6 +68,13 @@ function DeleteConsume() {
          .catch(err=>console.log("onRefresh readConsumeAll = > ", err))
     }
 
+    const tableTitle = [
+        <FontText style={styles.text}>날짜</FontText>,
+        <FontText style={styles.text}>카테고리</FontText>,
+        <FontText style={styles.text}>소비가격</FontText>,
+        <FontText style={styles.text}>내용</FontText>,
+        <FontText style={styles.text}>삭제</FontText>,
+    ]
 
     if(!data){
         return(
@@ -61,8 +85,8 @@ function DeleteConsume() {
     return (  
         <View style={styles.tableContainer}>
             <Table>
-                <Row data={["날짜","카테고리","소비가격","내용","삭제"]} flexArr={[1, 1, 1, 1, 1]} style={styles.head} textStyle={styles.text}/>
-                <Rows data={data} flexArr={[1, 1, 1, 1, 1]} style={styles.row} textStyle={styles.text} />
+                <Row data={tableTitle} flexArr={[1, 1, 1, 1, 1]} style={styles.head}/>
+                <Rows data={data} flexArr={[1, 1, 1, 1, 1]} style={styles.row} />
             </Table>
         </View>
     );
@@ -75,7 +99,7 @@ const styles = StyleSheet.create({
     },
 
     // 표
-    tableContainer: { marginBottom : 30,backgroundColor: '#fff' },
+    tableContainer: {marginBottom : 30,backgroundColor: '#fff' },
     head: {  height: 40,  backgroundColor: '#f1f8ff', borderWidth : 1  },
     row: {  height: 30 , borderLeftWidth :1, borderBottomWidth :1 , borderRightWidth : 1  },
     text: { textAlign: 'center' }
